Restore OS-specific import paths broken by #1023
PR #1023 (DB Abstraction Layer) renamed osbenchmark.worker_coordinator.runner
-> runners/opensearch.py and split base classes into runners/base.py. The
matching workloads PR (opensearch-project/opensearch-benchmark-workloads#764)
that updated workload imports to the new paths had to be reverted because
released-OSB integration tests pull a Docker image of the previous OSB version
and break against the new paths. As a result, upstream workloads on branch '3'
still import:

    from osbenchmark.worker_coordinator.runner import PutSettings, Retry, Runner
    from osbenchmark.client import RequestContextHolder

This change restores the historical OS-side layout so released-OSB workloads
keep working unchanged, while leaving the new database abstraction layer in
place for non-OpenSearch backends (Vespa, Milvus, ...).

OS-side restoration:
- Recreate osbenchmark/worker_coordinator/runner.py at its 2.1 path by
  combining runners/{base.py, opensearch.py, __init__.py} into a single
  module. Functionally identical to the post-#1011 split.
- Delete runners/{base.py, opensearch.py} and reduce runners/__init__.py to
  a minimal package marker so the runners/ subpackage exists only for new-DB
  modules (vespa.py, milvus.py).
- Revert osbenchmark/worker_coordinator/worker_coordinator.py imports of
  client and runner back to the canonical 2.1 paths. Restore the OpenSearch
  client construction in WorkerCoordinator and AsyncIoAdapter to be byte-
  identical to 2.1; non-OpenSearch backends are routed through the database
  abstraction layer in an explicit `if db_type != 'opensearch'` branch.
- Revert it/__init__.py, osbenchmark/builder/builder.py, and
  osbenchmark/workload_generator/workload_generator.py imports.
- Revert the same imports in tests/ and benchmarks/, plus the
  logger-name strings in tests/client_test.py.

De-duplication of database/clients/opensearch/opensearch.py:
- The 657-line file was a verbatim copy of osbenchmark/client.py plus the
  DatabaseClient adapter wrappers, and the two had already drifted (the
  lazy-grpc-import bug fix landed in only one of them). Replace the
  duplicated upper section with a re-export from osbenchmark.client. The
  adapter classes (OpenSearchClientFactory, namespace wrappers) stay
  unchanged. Single source of truth, no drift risk.

Net diff: +451 / -858 lines vs upstream/main. No behavior change for OS;
non-OS database routing is unchanged. All 1386 unit tests pass.
