+++ osbenchmark/worker_coordinator/runners/__init__.py
@@ -23,101 +23,13 @@
 # under the License.
 
 """
-Runner package for OpenSearch Benchmark.
+Subpackage for non-OpenSearch database runner implementations.
 
-This package contains:
-- base.py: Database-agnostic runner infrastructure (base classes, utilities)
-- opensearch.py: OpenSearch-specific runner implementations
-- vespa.py: Vespa runner implementations (stub for future PR)
+OpenSearch runners and the runner registry live at the historical path
+``osbenchmark.worker_coordinator.runner`` (singular) so that the workloads
+repository can keep importing OpenSearch runner classes unchanged.
 
-The registry functions (register_runner, runner_for, etc.) live here in __init__.py
-and are the central coordination point for all runner registration.
+New database backends (Vespa, Milvus, ...) live in this subpackage as
+sibling modules and import their base classes from
+``osbenchmark.worker_coordinator.runner``.
 """
-
-import logging
-import types
-
-from osbenchmark import exceptions, workload
-from osbenchmark.worker_coordinator.runners.base import (
-    Runner,
-    Delegator,
-    time_func,
-    request_context_holder,
-    mandatory,
-    escape,
-    remove_prefix,
-    unwrap,
-    _single_cluster_runner,
-    _multi_cluster_runner,
-    _with_assertions,
-    _with_completion,
-    MultiClientRunner,
-    AssertingRunner,
-    NoCompletion,
-    WithCompletion,
-)
-
-__RUNNERS = {}
-
-
-def register_runner(operation_type, runner, **kwargs):
-    logger = logging.getLogger(__name__)
-    async_runner = kwargs.get("async_runner", False)
-    if isinstance(operation_type, workload.OperationType):
-        operation_type = operation_type.to_hyphenated_string()
-
-    if not async_runner:
-        raise exceptions.BenchmarkAssertionError(
-            "Runner [{}] must be implemented as async runner and registered with async_runner=True.".format(str(runner)))
-
-    if getattr(runner, "multi_cluster", False):
-        if "__aenter__" in dir(runner) and "__aexit__" in dir(runner):
-            if logger.isEnabledFor(logging.DEBUG):
-                logger.debug("Registering runner object [%s] for [%s].", str(runner), str(operation_type))
-            cluster_aware_runner = _multi_cluster_runner(runner, str(runner), context_manager_enabled=True)
-        else:
-            if logger.isEnabledFor(logging.DEBUG):
-                logger.debug("Registering context-manager capable runner object [%s] for [%s].", str(runner), str(operation_type))
-            cluster_aware_runner = _multi_cluster_runner(runner, str(runner))
-    # we'd rather use callable() but this will erroneously also classify a class as callable...
-    elif isinstance(runner, types.FunctionType):
-        if logger.isEnabledFor(logging.DEBUG):
-            logger.debug("Registering runner function [%s] for [%s].", str(runner), str(operation_type))
-        cluster_aware_runner = _single_cluster_runner(runner, runner.__name__)
-    elif "__aenter__" in dir(runner) and "__aexit__" in dir(runner):
-        if logger.isEnabledFor(logging.DEBUG):
-            logger.debug("Registering context-manager capable runner object [%s] for [%s].", str(runner), str(operation_type))
-        cluster_aware_runner = _single_cluster_runner(runner, str(runner), context_manager_enabled=True)
-    else:
-        if logger.isEnabledFor(logging.DEBUG):
-            logger.debug("Registering runner object [%s] for [%s].", str(runner), str(operation_type))
-        cluster_aware_runner = _single_cluster_runner(runner, str(runner))
-
-    __RUNNERS[operation_type] = _with_completion(_with_assertions(cluster_aware_runner))
-
-
-def runner_for(operation_type):
-    try:
-        return __RUNNERS[operation_type]
-    except KeyError:
-        raise exceptions.BenchmarkError("No runner available for operation type [%s]" % operation_type)
-
-
-def enable_assertions(enabled):
-    """
-    Changes whether assertions are enabled. The status changes for all tasks that are executed after this call.
-
-    :param enabled: ``True`` to enable assertions, ``False`` to disable them.
-    """
-    AssertingRunner.assertions_enabled = enabled
-
-
-# Only intended for unit-testing!
-def remove_runner(operation_type):
-    del __RUNNERS[operation_type]
-
-
-# Re-export OpenSearch runners and register_default_runners for convenience.
-# These must come after the registry functions above since opensearch.py imports runner_for at module level.
-from osbenchmark.worker_coordinator.runners.opensearch import *  # noqa: F401,F403,E402  # pylint: disable=wrong-import-position
-from osbenchmark.worker_coordinator.runners.opensearch import register_default_runners  # noqa: E402  # pylint: disable=wrong-import-position
