+++ tests/kafka_client_test.py
@@ -25,7 +25,7 @@
 from unittest import TestCase, mock
 
 from osbenchmark.kafka_client import KafkaMessageProducer
-from osbenchmark.database.clients.opensearch.opensearch import MessageProducerFactory
+from osbenchmark.client import MessageProducerFactory
 from tests import run_async
